Update swift-quiz.md
Q30 was wrong. AnyObject cannot be a function or an optional. the code would not compile if I assign a closure to it and I get an error : `Value of type '() -> AnyObject' expected to be instance of class or class-constrained type`

Added the reference page from the apple doc. `Any` would be a correct answer...